+++ src/Symfony/Component/Security/Http/Authentication/AuthenticatorManager.php
@@ -62,12 +62,19 @@ class AuthenticatorManager implements AuthenticatorManagerInterface, UserAuthent
     }
 
     /**
-     * @param BadgeInterface[] $badges Optionally, pass some Passport badges to use for the manual login
+     * @param BadgeInterface[]     $badges     Optionally, pass some Passport badges to use for the manual login
+     * @param array<string, mixed> $attributes Optionally, pass some Passport attributes to use for the manual login
      */
-    public function authenticateUser(UserInterface $user, AuthenticatorInterface $authenticator, Request $request, array $badges = []): ?Response
+    public function authenticateUser(UserInterface $user, AuthenticatorInterface $authenticator, Request $request, array $badges = [] /* , array $attributes = [] */): ?Response
     {
+        $attributes = 4 < \func_num_args() ? func_get_arg(4) : [];
+
         // create an authentication token for the User
         $passport = new SelfValidatingPassport(new UserBadge($user->getUserIdentifier(), fn () => $user), $badges);
+        foreach ($attributes as $k => $v) {
+            $passport->setAttribute($k, $v);
+        }
+
         $token = $authenticator->createToken($passport, $this->firewallName);
 
         // announce the authentication token